# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MedineTech - ERP - Management Software API",
        "description": "Integration API Docs.",
        "contact": {
            "name": "Developers Team",
            "url": "https://medine.tech/",
            "email": "devs@medine.tech"
        },
        "version": "0.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Localhost Server"
        }
    ],
    "paths": {
...
        "/api/backoffice/{tenant}/accounting/accounting-accounts": {
            "get": {
                "tags": [
                    "Backoffice - Accounting - Accounting Accounts"
                ],
                "summary": "Search accounting accounts based on filters",
                "operationId": "50b5c7109b85e1930afa54a0b7651a53",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "The tenant ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of accounting accounts list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "123e4567-e89b-12d3-a456-426655440000"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "101"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Cash Account"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Main cash account",
                                                        "nullable": true
                                                    },
                                                    "type": {
                                                        "description": "1 = asset, 2 = liability, 3 = equity, 4 = revenue, 5 = expense",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "status": {
                                                        "description": "Account status",
                                                        "type": "string",
                                                        "example": "ACTIVE"
                                                    },
                                                    "parent_id": {
                                                        "type": "string",
                                                        "example": "123e4567-e89b-12d3-a456-426655440001",
                                                        "nullable": true
                                                    },
                                                    "creator_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "updater_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "company_id": {
                                                        "type": "string",
                                                        "example": "company-123"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "You do not have permission to view this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Backoffice - Accounting - Accounting Accounts"
                ],
                "summary": "Create a new accounting account",
                "operationId": "8fa28129042baf3e6f59635291d4d927",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "code",
                                    "type"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "123e4567-e89b-12d3-a456-426655440000"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Cash Account"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cash account for the company",
                                        "nullable": true
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "101"
                                    },
                                    "type": {
                                        "description": "1 = asset, 2 = liability, 3 = equity, 4 = revenue, 5 = expense",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "123e4567-e89b-12d3-a456-426655440001",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Accounting account created successfully"
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "You do not have permission to view this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "An unexpected error occurred while processing your request."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}